Done post boxchat

+++ src/main/java/com/example/chatting/app/repository/TokenRepository.java
@@ -15,5 +15,5 @@ public interface TokenRepository extends JpaRepository<Token,Long> {
       on t.user.id = u.id\s
       where u.id = :id and (t.expired = false or t.revoked = false)\s
       """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 }
